// src/enums/algorithm.ts
enum Algo {
    SUPERVISED = 'SUPERVISED',
    UNSUPERVISED = 'UNSUPERVISED',
    REINFORCEMENT = 'REINFORCEMENT',
    ENSEMBLE = 'ENSEMBLE',
    REGRESSION = 'REGRESSION',
    CLASSIFICATION = 'CLASSIFICATION',
    DECISION_TREE = 'DECISION TREE',
    CLUSTERING = 'CLUSTERING',
    K_MEANS = 'K-MEANS',
    DEEP_Q_LEARNING = 'DEEP Q-LEARNING',
    NEURAL_NETWORK = 'NEURAL NETWORK',
    VECTORIZATION = 'VECTORIZATION',
    SIMILARITY_SEARCH = 'SIMILARITY SEARCH',
    ISOLATION_FOREST = 'ISOLATION FOREST'
}

export default Algo